Allow creating gallery image versions from storage accounts (#1466)
* Allow creating gallery image versions from storage accounts

It's possible to upload a VHD into a container in a storage account and
then use that to import an image version. This adds a third way to
specify an `os_disk` in the galleryimageversion plugin to support that
approach.

* Address PR comments for style

There's no functional change.

* Fix example and tests based on review comments

Add the missing "source" intermediate key in the example and test case,
and ensure the VM information isn't shadowed by later steps.

* galleryimageversion: rename image in test created from URI

This also removes the image version at the end so the gallery can be
deleted.