Implement long running polling (#1649)
This implements long running polling the same way the GenericRestClient
does with a 600 second overall timeout and 30 second polling.

Implement retries and backoffs in the batch processing.  I looked at the
Azure SDK's pipeline which supports policies like RetryPolicy but I
don't think it will help here.  It doesn't look like we get a 429 when
sending the request to the /batch endpoint, only in the batch responses.